move mset out of msg pool file [skip ci]
